+++ packages/perseus-editor/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @khanacademy/perseus-editor
 
+## 4.2.8
+
+### Patch Changes
+
+-   Updated dependencies [[`b1552885`](https://github.com/Khan/perseus/commit/b155288568a3f13b3983515391cdbf4dd5092a51)]:
+    -   @khanacademy/perseus@19.6.1
+
+## 4.2.7
+
+### Patch Changes
+
+-   Updated dependencies [[`0562929c`](https://github.com/Khan/perseus/commit/0562929c03645028b68da497382489747f0fc2c6), [`4e2b5100`](https://github.com/Khan/perseus/commit/4e2b51002e41c8f36051775a7ceff4fd1a564526)]:
+    -   @khanacademy/perseus@19.6.0
+
+## 4.2.6
+
+### Patch Changes
+
+-   [#1020](https://github.com/Khan/perseus/pull/1020) [`30405816`](https://github.com/Khan/perseus/commit/3040581645a67f518920fdf6d94b4f78f2233f0c) Thanks [@nishasy](https://github.com/nishasy)! - Undo update to GraphSettings making it a class function
+
+-   Updated dependencies [[`30405816`](https://github.com/Khan/perseus/commit/3040581645a67f518920fdf6d94b4f78f2233f0c)]:
+    -   @khanacademy/perseus@19.5.1
+
 ## 4.2.5
 
 ### Patch Changes